backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.1
numpy==2.2.6
statsmodels==0.14.5
python-dotenv==1.1.1
httpx==0.28.1
pydantic==2.11.7
